+++ src/Frontend/Modules/Faq/Js/Faq.js
@@ -4,6 +4,7 @@
  * @author	Annelies Van Extergem <annelies@netlash.com>
  * @author	Jelmer Snoeck <jelmer@siphoc.com>
  * @author	Thomas Deceuninck <thomas@fronto.be>
+ * @author	Jeroen Desloovere <jeroen@siesqo.be>
  */
 jsFrontend.faq =
 {
@@ -22,17 +23,19 @@ jsFrontend.faq.feedback =
 		// useful status has been changed
 		$('#usefulY, #usefulN').on('click', function()
 		{
-			// get useful status
-			var useful = ($('#usefulY').attr('checked') ? true : false);
+			// init useful status
+			var useful = false;
+
+			if ($(this).attr('id') == 'usefulY') {
+				// get useful status
+				useful = ($("#faqFeedbackForm input[type='radio']:checked").val() === 'Y');
+			}
 
 			// show or hide the form
-			if(useful)
-			{
+			if (useful) {
 				$('#message').prop('required', false);
 				$('form#feedback').submit();
-			}
-			else
-			{
+			} else {
 				$('#feedbackNoInfo').show();
 				$('#message').prop('required', true);
 			}